fix: 리뷰 개수 필드의 타입을 int로 수정

+++ src/main/java/com/codeit/mission/deokhugam/book/dto/BookDto.java
@@ -14,7 +14,7 @@ public record BookDto(
     LocalDate publishedDate,
     String isbn,
     String thumbnailUrl,
-    long reviewCount,
+    int reviewCount,
     BigDecimal rating,
     LocalDateTime createdAt,
     LocalDateTime updatedAt
